Drop in styling from the codemash html repo
Since angular 2 lets us style against components directly, we can use less specific selectors. The only question will be which styles should be applied globally and which should be applied to specific components. For the header in particular, this allows us to style directly against the header without affecting the app styling.