Show the events in the category
For now we'll move the styling to the global file. We'll have to determine where to put these after we figure out what components need to make use of these. It's possible we could put these in shared styling files that get imported into all of the components that use it. (Components have an array of style urls, so this could work nicely)