Created an http service to encapsulate the details of making http calls
Things like content headers, parsing, and formatting json will need to be done by everyone who makes an http call, so we might as well put these into a reusable service. This makes http calls much simpler in the consuming components. Also, per recommendation of the docs, it's best to declare all 'global' providers on the app module, even if they are explicitly used in a lazy-loaded submodule. Otherwise the lazy loaded module will create its own instance of the service. (Shouldn't hurt for the http service, but we'll leave it open in case the services module gets other services that do need to be completely global.